axxiaarm: u-boot-fw-utils: add HOSTCC to OEMAKE variable
Issue: LIN10-3998

Since "/usr/bin" is not in WRL10's $PATH environment variables anymore, when
compile u-boot-fw-utils which based on u-boot-axxia, a compile error will be
triggered as below because command "cc" can't be found:

/bin/sh: cc: command not found

So add the definition of HOSTCC to the OEMAKE variable to avoid this error.
But when HOSTCC is set with a command argument, the ordinary assigments in makefile
are ignored, this means "HOSTCC = $(CC)" in tools/env/Makefile will not work, and
tools/env can't be cross-compiled anymore. Pickup a uboot upstream patch to fix it.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>