axxiaarm: install bl31.o to u-boot-axxia's sysroot dir
Issue: LIN1021-1225

When building project with SSTATE_DIR, there is an occasion that
atf-axxia won't run do_compile().

So rewrite bb file of atf-axxia and u-boot-axxia to avoid this:
1) Move the operation of installing bl31.o from do_compile() to
do_install().
1) And for now bl31.o is installed to ${STAGING_DIR} which is
tmp-glibc/sysroots. This is not quite good and it's better to
package it and install it to the sysroot dir of u-boot-axxia.
This is to ensure that the dependency between u-boot-axxia and
atf-axxia works.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>