axxia: bump up u-boot version for axm55xx to 5.8.1.90
Issue: LIN1021-1707

Since the new u-boot 5.8.1.90 contains CVE2017-571 solution, let's
switch to this newer version.

Modifications as below:
1. delete obsoloste patches which is for older u-boot.
2. add a new patch to fix the issue in the newer version u-boot.
3. switch u-boot git repo for axm55xx from
    git://github.com/lsigithub/lsi_axxia_uboot_public.git
to
    git://github.com/axxia/axxia_u-boot.git
which is also the repo of axm56xx's u-boot. Since the new repo has
some IP issue, delete it from PNWHITELIST and require the customers
to download it by themselves.
4. update description in README and README.hardware for the newer
version.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>